RDMA/hbl: add habanalabs RDMA driver
Add an RDMA driver of Gaudi ASICs family for AI scaling.
The driver itself is agnostic to the ASIC in action, it operates according
to the capabilities that were passed on device initialization.
The device is initialized by the hbl_cn driver via auxiliary bus.
The driver also supports QP resource tracking and port/device HW counters.

Signed-off-by: Omer Shpigelman <oshpigelman@habana.ai>
Signed-off-by: Abhilash K V <kvabhilash@habana.ai>
Signed-off-by: Andrey Agranovich <aagranovich@habana.ai>
Signed-off-by: Bharat Jauhari <bjauhari@habana.ai>
Signed-off-by: David Meriin <dmeriin@habana.ai>
Signed-off-by: Sagiv Ozeri <sozeri@habana.ai>
Signed-off-by: Zvika Yehudai <zyehudai@habana.ai>